Move `Iteration` from 'lib.rs' into separate file 'iteration.rs'
